requirements.txt
blis==0.7.10
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
confection==0.1.0
cymem==2.0.7
dnspython==2.4.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.6.0/en_core_web_lg-3.6.0-py3-none-any.whl
filelock==3.12.2
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.3
murmurhash==1.0.9
numpy==1.25.1
packaging==23.1
pandas==2.0.3
pathy==0.10.2
phonenumbers==8.13.17
preshed==3.0.8
presidio-analyzer==2.2.33
pydantic==1.10.12
pymongo==4.4.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
six==1.16.0
smart-open==6.3.0
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
thinc==8.1.10
tldextract==3.4.4
tqdm==4.65.0
typer==0.9.0
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wasabi==1.1.2
jsonpathgenerator
jsonpath-ng
